requirements.txt
backports-abc==0.5
bkcharts==0.2
bokeh==0.12.5
certifi==2017.4.17
chardet==3.0.4
click==6.7
Flask==0.12.2
futures==3.1.1
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
numpy==1.13.0
paho-mqtt==1.2.3
pandas==0.20.2
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
requests==2.18.1
rethinkdb==2.3.0.post6
singledispatch==3.4.0.3
six==1.10.0
tornado==4.5.1
urllib3==1.21.1
Werkzeug==0.12.2
